Remove straight_join in favor of index hints
The build queries that straight_join improves are starting to look
a lot like the set that are improved by the existing index hint,
so remove the use of straight_join there and rely only on the hint.

A similar characteristic holds for the buildset queries.  We may
be able to use index hints for those as well, but to get good
performance, we need to set the hint on two tables.

Change-Id: Idf622fb1ecbf933b4570dd35b4c3cf4750e69c75