Always store zstat for sharded zkobjects
Traditionally a sharded zkobject did not store a zstat because
it isn't represented by a single object.  But some users or
subclasses do use a zstat of a sharded zkobject.  They rely on
the fact that we always truncate the contents of the object before
writing, and truncating means removing the base znode.  That means
that each time we write, we will create a new base znode, and that
is sufficient to act as a zstat for the whole object.

This change normalizes this and implements it in the ShardedZKObject
class, and removes some now duplicated code.

Change-Id: I4b973dcfb1175e882f7e3c33d61867424dd3f654