Add some of the aws driver and statemachine framework
This is part of nodepool in zuul.

This imports most of the aws adapter implementation from nodepool,
along with the fake_aws class and some test framework to make it
work.  It also imports some of the statemachine base classes,
but not the statemachine engine itself.

It updates the aws driver to start and stop the threads/executors
on the adapter.

Changes so far are fairly minimal, and much of this code will not
actually execute due to required changes.  It is being ported over
mostly as-is so that later changes that actually change the
implementation compared to the current system are reviewable as
deltas.

The only significant changes in this commit are:
* The QuotaInformation class has been flattened so that it does not
  contain a "resources" dictionary.  All resources (instances, cores,
  etc) will be at the top level.
* A new test decorator for driver configuration has been added.  This
  will let us add test config options for any driver, this way we
  don't need to keep adding individual test decorators.  We should
  probably remove @gerrit_config and update it's use to this.

One notable non-change: the method of overriding the connection/adapter
implementation in the aws tests remains the same as in nodepool: we
swap out the class implementation in the test rather than overriding
the getConnection method as most other zuul drivers do.  If we can
sustain this, this will make for fewer deltas between test and production
code.

Change-Id: Ib478096dc7c866efd7b6a8445cd0d8751b105eb2