WIP: demo compressed non-live items
In review comments we have discussed the presentation of non-live
items in independent pipelines in the overview page.  These are
currently rendered as grey squares, and a skilled user can read
the tile field in an independent pipeline as follows: a series of X
grey squares preceding a non-grey square indicates the non-grey
square depends on a series of X changes.

However, grey is also used elsewhere to mean "this live queue item
has not started yet"; that could make things a little confusing.

The useful pieces of information are:
* How many dependencies the change has (operators want to know how
  much extra work is being done to test each change).
* What actual changes are being tested with a change (users want to
  see what's being tested with their change).

Inasmuch as the current overview page allows users to hover or click
on a tile, that satisfies the second condition (though with some extra
effort to hover/click on each in order to identify the change).
Arguably, the pipeline detail view is a better way to do that since
all dependent queue items can be seen at once.

If we sacrifice (somewhat) the ability to hover over the non-live
changes, we can present the information in item #1 (how many dependent
changes are included) in an easier to read manner by rendering the
non-live changes as small vertical bars.  We wouldn't expect users to
actually count the bars (though they could), but rather they can see
at a glance: here is a change with no dependencies; here is a change
with one or a small number; and here is a change with way too many
dependencies -- perhaps that one should be investigated.

One way to look at independent pipelines is that they are
two-dimensional. This can easily be seen on the pipeline detail view,
where the width is the number of live items, and the height of each
is the number of changes in them.  This brings an aspect of that
two-dimensionality to the overview page.

(Note: another way to do this would be to make that two-dimensionality
much more literal by rotating what you see in this change 90 degrees
so that each tile has a height corresponding to its dependency depth.
That might be interesting to explore, but it comes with more layout
challenges.)

This change is not how we should actually implement this; this is a
quick hack using the pipe character.  We should actually find an
appropriate icon or use better css styling to draw the lines.

Change-Id: I608daec4f4c5463490b82ffc584500814cfe9f6f