Add listResources/deleteResource from endpoint to provider
This was originally on the Adapter class in nodepool, but due
to the split in Zuul, it makes more sense to be on the Provider
rather than the Endpoint.

This is because when listing resources, we may need to consult
provider-specific information.  An example here is the bucket name
for image uploads.  That is configured on the provider, not the
connection.

The provider will use the endpoint methods to accomplish its
tasks.

Change-Id: I89aab1fcc5e33e3e1d1021e2224a671527711a2f