Pass provider_name to _tag methods
This adjusts the _tag* methods on the aws endpoint class to accept
a provider name as input.  These methods attempt to ensure that all
the resources that shoud be tagged are, but they are intended only
to deal with resources created by that provider.  In nodepool,
they checked the provider_name metadata.  We need to consider what
to use in Zuul:

* endpoint name: This is an option but only since, at the moment,
  we don't need any provider-specific information (because we copy
  the tag values from other already-existing objects).  But the
  objects themselves are created by a provider and may have provider-
  specific information, so it may make more sense to consider this
  a provider-scoped operation.
* provider canonical_name: This would mean that every
  tenant-instantiated provider would examine the objects for ever
  other instance of the same canonical provider.  This would work,
  but seems like undesirable overlap.  Thit would be a good choice
  if we moved the periodic maintenance of leaked objects to operate
  on the canonical provider level rather than the tenant-instantiated
  provider level, but we don't have a facility for tha yet.
* tenant+provider: This is the closest to what nodepool does today
  and does not have any overlap (though it performs a little extra
  work to skip objects from other instances of the same canonical
  provider).

This change implements the third option, and to make that simpler,
introduces a tenant_scoped_name attribute of the provider so that
we can consistently refer to the tenant-instantiated providers.

Change-Id: Ia26f5228677f4ba8a7bb2f2955eea46a9484f802