Move some attributes from label to flavor
The mapping for the kind of instance type to use is a flavor attribute,
so add flavor support to the base provider and the aws provider and
move the label attributes there.

Also make the provider attributes of flavors/labels/images into
dicts so that we can easily look up the references between them
later when we need to find out what instance type to create for
a given label.

Change-Id: Iea564af8eb8bc3fa10f06b48bd6ffdb632618e72