Split awsprovider.py
This moves the endpoint and related classes into their own file,
awsendpoint.py.  This looks a lot like the old adapter file and
is very implementation specific.

The constants are moved to their own file since they are getting
rather large.

The remainder in awsprovider.py is code that maps between the
concept of a zuul provider and the endpoint implementation.

Change-Id: I507301eb0323de95fc3f9b740b1aa375a9e83094