WIP: Drive statemachine in launcher
This adds some skeleton code to create a state machine in the
launcher and drive it.  A number of features in the AWS driver
still need configuration settings added, so they have been commented
out with "TODO" comments to add them back in as the configuration
matures.

This adds stub attributes to the node to store the state machine
data, but does not yet serialize it to ZK.

Change-Id: I6f2404afddcc5373016bcadefb3cb97857380a16