Deprecate and remove usage of web.status-url
It seems that there are different ways to configure the zuul-web root
URL and also inconsistent usage.

The the web root/status-url is part of the system attributes in
Zookeeper. This means we must not get those config options from the
local config file.

Additionally, the status-url can be derived from the web root for the
global web root URL as well as when using a "whitelabel tenant" web-root
setting. To avoid confusion the web.status-url setting is removed.

This change also removes the status-url setting from the Github and
Pagure reporter config.

Change-Id: If4ed1d7942dd2e7c2351a1eba9892f8503990f1f