Deprecate and cleanup use of status-url
In addition to the required web.root setting there is the web.status-url
that might cause some confusion for users as the difference is not
immediately clear.

In order to deprecate the web.status-url setting the following cases
need to be considered:

1. The Github reporter allows configuring the status-url as part of the
   pipeline configuration, but already uses the better item URL in
   `setCommitStatus()` instead. Here we can simply add a deprecation
   warning in case the status-url is configured.
2. The Pagure reporter also allows to configure the status-url and so
   far used it also setting the commit status. Analogous to the Github
   reporter we can replace this with the item URL and also add a
   deprecation warning when it is configured.
3. The no-job, enqueue and start messages still use the global
   `status_url` default (now unified to always use the setting from the
   system attributes and NOT from the local Zuul config). Since the
   `status_url` replacement attribute has already been deprecated in
   favor of `item_url` there isn't anything to do here.

After we've removed support for the `status_url` format attribute, we
can also remove the global `web.status_url` setting from the config and
the system attributes in Zookeeper.

Change-Id: If4ed1d7942dd2e7c2351a1eba9892f8503990f1f