Log error details during tenant validation
So far we will only log the representation of ConfigErrors during tenant
validation, which is not useful:

    2024-07-19 15:21:16,233 WARNING zuul.Scheduler: <zuul.model.ConfigurationError object at 0x7f7fd01bba90>
    2024-07-19 15:21:16,233 WARNING zuul.Scheduler: <zuul.model.ConfigurationError object at 0x7f7ee1dc3450>
    ...

Instead we will use the full detail of all error messages when failing
the tenant validation. The warning messages are still logged in the
configloader.

We need to include the full error details as the first ten config
errors/warnings logged in the configloader might only be warnings.

Change-Id: I2e6f9df3ebea54ca5110108a6c9a9537856d9f1b