WIP: Unify provider validation and deserialization schemas
We are going to add a lot of configuration options related to
providers, and there will be some duplication.  Each Section
may have a number of attributes that will act as defaults for
labels, flavors, and images.  Labels, Flavors, and Images may
set those same values.  Providers may override them.

We need to validate each of these values in each of those places.
Then we also need to serialize and deserialize them to share
between the schedulers and launchers.

The more modular and reusable the schemas are, the easier it
will be to understand and maintain this.  To that end, this
change introduces a method of unifying the validation and
deserialization schemas.  This will allow us to use the same
voluptuous schema to validate the YAML as well as set the
attributes on the python object.

It allows mutating the key names so that "foo-bar" in YAML is
translated to "foo_bar" in python.  It also allows us to
set the python attribute to None if no value is provided without
also allowing None in the YAML input.

Change-Id: I11932d2d7da8ecd26010319db5be04c19d64756f