Serialize state machine state to Zookeeper
With this change we will move the state machine create/delete state into
separate subclasses. The state will be serialized as part of the
provider node.

To make sure that the SM state can't be modified, a ZKObjectMember
subclass is introduced that will prevent modifications of the state
without an active context in the parent ZKObject instance.

Provider methods for getting the create/delete state machine instances
have been simplified to only require a provider node. The state machine
will take care of creating a new or re-storing the existing SM state.

Change-Id: I7cca13c6991a1a7ea3508c0814cc57d04d76877d