Refactor AWS SMs to allow recovery from ZK state
This changes refactors the AWS create/delete state machine logic, so we
can more easily recover from the saved SM state in Zookeeper.

This means that we only store recoverable steps as part of the state.
With that we also no longer store separate start/submit host and
instance create states, but use an idempotency token for the API
requests (client token) so that we can re-try the calls if needed.

https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html

Change-Id: I182e24f1d78f6a660d1fb966f0e68690dea11fb5