WIP: Make QueueItem cards in ChangeQueue more concise
This is a follow-up on change 923778[1] where we try to integrate the
ChangeQueue from the PipelineDetails view into the Status view via an
expandable section.

This change removes some of the paddings used within the custom
zuul-compact-card or makes them smaller. Additionally, use a larger
font-size for the pipeline name, so it better stands out from the queue
names.

Additional changes are made to the style of the progress stepper to make
the icons smaller and to better align the progress stepper icons with
the content (QueueItem cards).

Finally, the AngleRight and AngleDown icons (to open/close the
expandable section) are aligned with the queue names.

Originally, I wanted to apply this change only to the ChangeQueue as
visualized on the Status view, but maybe we can keep it also for the
PipelineDetails view (currently it's applied to both).

[1]: https://review.opendev.org/c/zuul/zuul/+/923778

Change-Id: I17eb8203da78aa80116ebd95679e06b05fc0e434