Mark some intentional errors as okay
This marks several tracebacks as okay; these are generally instances
where we are intentionally causing errors in tests.

A new default for the okay_tracebacks value is added that matches
cases where we log an exception for a git merge conflict.  This happens
somewhat frequently in tests, intentionally, and should be safe to
ignore in all cases.  Moreover, this reduces future friction when adding
new tests that have git merge conflicts.

Change-Id: I35fe020abdfe8337287dbf522b20b3491a1b594a