Avoid DB exceptions on certain build cleanups
If we cancel a build very early (before it starts) and there have
been no builds started for the buildset, and the buildset wasn't
already reported as started to the DB, then in that case we may
log an exception when cleaning up the build after the cancelation.

We always attempt to report a build end to the DB, but in this case
we don't have any existing DB records, which means we first need
to create a buildset entry in the DB.  But we don't have enough
information to do that properly, so we throw an exception.  This
exception is caught and logged, but it's not a completely unexpected
situation.  To avoid logging the spurious exception, indicate to the
lower levels that it's okay if we don't have a buildset and we should
just return without any further DB work.

Change-Id: I1d600255e0218e2a6c73d134be23da665f250165