Fix test_crd_gate_unknown
This test was missing some necessary setup, which caused an exception
to be raised from the test fake during pipeline processing.

The scenario is that change A depends on change B in a project that is
unknown to Zuul.  The result of the exception was nearly the correct
behavior, in that change A did not run any jobs, but we do actually
report an error message on change A indicating that the reason we
ran no jobs is because we can't work with change B.

By correcting the exception here, we now exercise the correct behavior
in the test.  Its assertions are updated to match.

Change-Id: Ic98daaac830f75e612d8bf22cd2379f3807b8c2f