Fix traceback in test_nodepool_pipeline_priority
This test uses a ref-updated event to trigger a specific pipeline,
but it did not merge the fake change before emitting the event.
This causes an error when trying to set the repo state, which logs
an exception.

The results were ignored since the test is only interested in the
nodepool behavior.  An additional build history check is added
for good measure.

Change-Id: Iecfa840455d49c0b0e294330e986e125bac55b46