Serialize state machine state to Zookeeper
With this change the state machine state will be serialized as part of
the provider node.

The host/instance submit states have been removed as they can't be
recovered individually. If the state machine needs to be re-created
while a create future is still pending, we'll try to reconcile the state
by checking if an instance matching our tags already exists.

However, since this works based on a cached list of instances, there is
still a small chance that more than one instance is created for a
provider node. We'll rely on the cleanup routine to terminate the extra
instance later on.

Change-Id: I7435460a7f9473e80672753fb2020d4fa5f6263b