Don't start the scheduler when validating tenants
The start method starts extra threads like the apscheduler for
sending stats, etc.  We don't want any of that running when we
are just validating tenants, so move the KeyStorage into the
initializer and avoid running start/stop when we are validating
tenants.

(Tracebacks from periodic stats jobs were observed in long test
runs.)

Change-Id: I1aebdc3b5e069ab8bf8044de9d1121ae8d9798d7