Improve tracing test isolation
We never ran the cleanup method for the OTLP fake server, which
means we never shut it down properly.  This would leak threadpool
executor threads.

We also need to reset the tracing configuration at the end of tests,
otherwise subsequent tests will end up using the tracing provider
from the previous test.  This is because we currently only run
the reset at the start of a test that using tracing; so new tests
that startup without using tracing won't run the provider reset.

This change corrects both issues, which should mean that the tracing
configuration is isolated to the currently running test.

Change-Id: Ib1edf51ad8ef3550b16dc888883f86a96bfb2dc2