Fix web dequeue of ref-triggered items
This restores a code branch which was introduced in 867177 [1], but
later removed in 913390 [2]. In the current state, ref-triggerd items
(e.g. from a periodic pipeline) can now longer be dequeued as the regex
does not match a ref like 'refs/heads/master'.

This change restores the old behaviour while keep the current
functionality.

[1]: https://review.opendev.org/c/zuul/zuul/+/867177
[2]: https://review.opendev.org/c/zuul/zuul/+/913390

Change-Id: I294aa734ebd0ee5cbc53874c63436db1c85eddd4