Combine reserved variable name lists in  check_varnames
There are two different sections of checking for reserved variable
names in this one function, with slightly different error messages.
Neither message is specifically informative about the name restriction
in the error message, so there's not a great reason to do this
two different ways.

Add the first set of words into the list and deal with it all the
same way.

Change-Id: I51327f0abf160a93e81aed46806c3f56ee2c9316