Fix wrong blob cleanup for items w/o job graph
The method for getting referenced blobs of an item did not return any
blob keys when the buildset did not have a job graph yet.

If there is a blob store cleanup between setting the merge result and
building the job graph. This can for example happen when the merge state
was set, but we abort pipeline processing early due to a pending
reconfiguration.

Change-Id: I9c3160c29b6396fad4e9d6c319660f684379568d