Reconcile merger and executor retry methods
The base merger class has a ZK retry method, and the executor class
has a similar method.  The base merger method was, incongruously,
checking the _running flag on the executor.  This means whenever it
ran on an executor, it worked correctly, but if it ran on a standalone
merger, if it encountered any retryable failures, it would raise an
exception due to the missing attribute.

To resolve this, we will combine the two methods (a long-standing TODO
item), and ensure the new method checks whichever running flag is
appropriate.

(Note that if an executor is configured not to act as a merger, then
the _merger_running flag will always be False, so we check the
_running flag on the executor class, and the _merger_running flag
on the BaseMergeServer class.)

Change-Id: I39476c71fabfca62a3b7074f989678891e037184
Co-Authored-By: Simon Westphahl <simon.westphahl@bmw.de>