Remove unused process_worker args
The getRepo and _addProject methods accept but do not use a
process_worker argument.  In particular, _addProject should not
accept a process worker because it should only be used for the
initial addition of a new repo to the merger, and that should
never happen in a process worker (since the resulting git.Repo
object would not actually be available to the main process).

To improve clarity and reduce the risk of future mistakes,
remove these arguments.

Change-Id: I234b9220a81f5daf2136f5e0e55ca9888498bb86