Be explicit about git.Repo lifecycle
The git.Repo object creates a pair of long-running "git cat-file" processes
which it uses to read data from the git repo.  Most operations that involve
commits trigger the start of these processes.  They are terminated when
the git.Repo object is gc'd, or when it is "closed", where "closed" may
mean an explicit call to the close() method, or via the exit call when
used as a context manager.

To improve the likelihood that these processes are actually terminated and
not leaked, this change does the following:

1) Turns almost all usages of the git.Repo object into a context manager.
   (The exceptions are a couple of calls that are part of the initialization
   process and don't involve any commits).
2) Avoids returning commit objects from methods, instead returning the
   string hexshas of those commits.  All of the instances where we used
   a returned commit can be simplified to just use the hexsha.

Some unused methods are cleaned up as well.

Change-Id: I6cfac53f86498f26dc906ac35cccc18aa0e22699