Don't log tracebacks on pipeline change list refresh
This situation, where a scheduler is unable to refresh the list of
changes in a pipeline because another one is writing that list
is expected.  We don't lock these reads because it's almost always
sufficient to use our last values in the cases where it happens.

The error is pretty much always exactly one thing: a JSON decode
error.  So if we encounter that one, we'll just log it as a warning.
If it happens for some other reason, we'll log the exception.

Change-Id: Idd92804ccf70519f71efe01129af63c77a67a97d