Inject queue_item_id and enqueue_time to inventory zuul.items
The use case for these variables is to perform repository updates
outside the internal Zuul merger. For that, all changes in the same
queue item need to be grouped, in the same way as the deprecated
bundle_id does. The enqueue_time is used as a timestamp for the git
commits, just as the Zuul merger does internally.

Change-Id: Ib0eb2df1268d0e3a8dbb6314b9da9babc9399be0