Add a timeout for gerrit ssh commands
We currently set a timeout for Gerrit HTTP requests, in case there
is a network problem or the Gerrit server hangs, or a proxy server
connection hangs.  This is set to 30 seconds because we don't expect
any individual query or merge action to take that long.

This change adds the same functionality to Gerrit SSH requests.  We
expect more reliable network connections with these, but it's still
possible for Gerrit to hang on a request.  In this case, the thread
in Gerrit will be lost until an administrator deals with it.  But
that's not Zuul's concern.  Zuul should attempt to recover from this
situation, even if it means using a new Gerrit thread.

Change-Id: I4093f0ddbe3b5daf04cf48c04a25872d00ec8e4d