Handle zuul_return format errors
This change does two things:

First, it corrects an error in the zuul_return schema validation
which would, in a job where someone returned an artifact with a
missing required attribute such as "url", cause an exception in
the SQL reporter (which would lead to incomplete build results in
the db, and possibly other pipeline-related errors).

This change is made by correcting the voluptuous schema so that
the "Required" flag is on the key, not the value.

Secondly, it performs schema validation within the zuul_return
module itself.  This will cause such errors to become user-visible
at the time they occur (during the job) rather than post-facto.
To accomplish this, the schema definition is moved into the
shared zuul/ansible module, and a dependency on voluptuous within
the in-job ansible virtualenvs is added.

The first fix is still needed even with the second because users
can return data to Zuul without using the zuul_return plugin.

Change-Id: I6991863b2b5986b8067124b92c141c3c59240d18