Add image-build-name job attribute
This adds the image-build-name job attribute which is used to decide
what jobs to run when an image is required, and will be used to populate
the internal image registry.

In order to avoid issues with unauthorized tenants building centrally
defined images, jobs with an image-build-name may only be attached to
project stanzas in the same project where the image they reference is
defined.

Change-Id: I16267fac6a867989a5412b014f97ae839cfdd55a