Add image upload records
This is part of nodepool in zuul.

This adds image upload records to ZK.  They are created after an
image build completes and will later (not in this change) prompt
the launchers to upload the image artifacts.

This expands (but does not complete) support for validation.  This
adds the ability to configure validation jobs, but they won't be
launched until after uploads occur, which is not supported yet.

Change-Id: Ibae3e3cb5bcbef7751c366c1298e026109cacea1