Be more explicit about quickstart container deps
Running podman 5.2.0 with podman-compose 1.2.0 on OpenSUSE Tumbleweed
I've run into problems with the executor container not starting during
quickstart. There are complaints in the logs about dependencies not
being resolvable and not being able to restart a container in an unknown
state.

Everything seems to work if I restart things later. Based on the errors
and things starting up later just fine my hunch is that podman and/or
podman-compose is racing container startup in some situations. If I
explicitly list mysql and zk as dependencies for the executor this goes
away. I listed these two as the executor depends on the scheduler which
depends on these two. I suspect that maybe transitive deps aren't
handled properly?

While at it I update the web container to depend on zk as it has a
direct dependency on the zk service too.

Change-Id: I320c1f46cad5806ac08e1f94d78a7e974cbb0e42