Remove unneeded depends_on from Zuul quickstart containers
Running podman 5.2.0 with podman-compose 1.2.0 on OpenSUSE Tumbleweed
I've run into problems with the executor container not starting during
quickstart. There are complaints in the logs about dependencies not
being resolvable and not being able to restart a container in an unknown
state. Everything seems to work if I restart things later. Based on the
errors and things starting up later just fine my hunch is that podman
and/or podman-compose is racing container startup in some situations and
is unable to handle transitive dependencies in particular.

It turns out that listing these dependencies appears to be unnecessary
as Zuul components will gracefully handle the loss and return of their
network dependencies like the database and zookeeper. Simplify the
quickstart by removing these depends_on entirely.

Change-Id: I320c1f46cad5806ac08e1f94d78a7e974cbb0e42