Drop use of sudo podman-compose from quickstart tutorials
The main Zuul quickstart tutorial does not use sudo to run
podman-compose. Instead all of the containers are run as your normal
user with podman. The sub tutorials for keycloak and jaeger do use sudo
to execute podman-compose.

This is probably fine from a functionality perspective as the services
all interact with one another over the network. That said there isn't
any reason to run some services as your normal user and others as root
when using podman. Remove the use of sudo to simplify things.

Change-Id: Id6fef5205b1c9208f5fab65b90000188298fac63