Implement tile scrubbing in status overview
The status popup currently only appears when clicked.  We have
experimented with having it appear as a hover effect, but there are
usability problems with that.

As an alternative, this allows the user to select that behavior
by "scrubbing".  If they click the pointer and drag it across the
queue item tiles, the popups will appear for each tile as the pointer
moves across them.  On release, the popup for the last tile will
remain and the behavior reverts to the normal click-to-open or -close.

This facilitates the use case of being able to quickly scan a subset
of queue items to identify commonalities (such as project) without
needing to leave or expand the overview.  It also keeps the popup
out of the way when it's not needed.

Change-Id: I2ceec401fef07be3eb93bcc23c1d3d06e8cb9c5d
Co-Authored-By: Felix Edel <felix.edel@bmw.de>