Implement validation reporter
This completes the (normal) validation cycle.  When uploads are
complete, image-validation events are emitted and when their
corresponding jobs complete successfully, the zookeeper image
upload record is updated.

The actual mechanism to substitute the new image in the nodeset
is not present in this change; that will need to wait for
the ability to launch nodes with external_image_ids.

Change-Id: I6b1e24d16bd4369ec48771cf68a599fd7a6035f0