Hide red cross icon next to progress bar for pre-fail jobs
When the pre-fail flag on a job is true, we apply the "danger" variant
to the job's progress bar. Apart from adapting the colors to visualize
the failure, this variant also adds a red cross icon next to the bar to
emphasize this.

This icon doesn't provide much benefit and breaks the layout of the job
progress bar. As I didn't find a way to remove it, let's just hide it
using CSS.

Before/after: https://imgur.com/a/fpWKSUJ

Change-Id: I2fa43f470d93134d63f9ef8592e679bab279f014