Use queue.name to differentiate between different queue visualizations
On the new status page we want to visualize named and unnamed queues
differently. Named queues will be visualized independently, while
unnamed queues will be squashed together in a single queue to simplify
the visualization.

So far we were using the pipeline type to differentiate between the
different render() code paths, but actually it's much easier to use the
queue.name attribute directly for the differentiation.

Additionally, use a nth-child ccs rule to avoid setting different
paddings based on the pipeline type.

Change-Id: Ib2f2df925b71a2bf3c39f4093f3a3943502250d8