Overwrite job result label colors to match the old status page design
On the old status page, the status of a job was represented by a single
box containing either a progress bar (if the job was running) or a label
with the result of the build. This worked well, because in Patternfly 3
progress bar and label were styled in a similar way which ensured a
seamless visual transition from progress to result.

To keep the same visual language on the new status page, we have to
align the colors of progress bar and label as both components are styled
quite differently in Patternfly 4.

Change-Id: I36b55555f24c8d006396c35fef7f7b5661721e46