WIP: Prevent pipeline cards from flickering
When an item shortly enters a pipeline but gets removed again
immediately, a pipeline card shortly appears and disappears, shuffling
the view unnecessarily. Those items might be flagged as live, but don't
have any jobs, as this happens before job freezing.

To prevent this flickering, hide empty pipelines not based on the number
of items, but on the sum of the jobs.

A side effect of this change is, that we first display pipelines a bit
later when the "Show all pipelines" toggle is off (only after job
freezing is done).

Change-Id: I8e6ba82876f2e920e6f10f6742abc5113a71edc5