Show window size and queue length per queue on new status page
The window size is a useful information that was so far not shown on
the new status page. This changes adds a badge next to the queue name to
show this information - similar to the old status page.

The badge is only shown for for named queues as those are visualized
independently. Unnamed queues are squashed together into a single queue
to simplify the visualization. For those it should be sufficient to show
the pipeline length.

Change-Id: Ic359c49f5d18b870de3f9b5793f9c631bc2d0bd2