Use AsleepIcon for non-live items
So far both, non-live and queued items, used the same OutlinedClockIcon.
To better differentiate between the two states, use an AsleepIcon for
non-live items instead.

The old status page used an InfoIcon for non-live items, but I find this
to prominent compared to the other icons we are using and IMHO it tends
to raise once attention although in the end it's only a non-live item.

Change-Id: Ib49a11f74b6949b3cf2de085ee4fc2d3715ad18c