Use InfoIcon for non-live items
So far both, non-live and queued items, used the same OutlinedClockIcon.
To better differentiate between the two states, use an InfoIcon for
non-live items instead.

In order to avoid raising undue attention, the info icon uses the "disabled"
color (typically a drak grey).  The progress stepper did not use the
colors we configured for the icons (only the build result pages), so
update the stepper to set the same icon colors as the build results.

Change-Id: Ia6de773fca003869f02ef207b87aed0b27d91f27
Co-Authored-By: James E. Blair <jim@acmegating.com>