Add filtering action to QueueItem
This adds different filter actions to a QueueItem. The actions are
reachable via a search-plus icon which will open a dropdown to present
a selection of different filters.

Selecting a filter will override the current URL params (and thus the
currently applied filters). However, it is still possible to apply
additional filters from the toolbar after selecting a filter from the
dropdown.

Using a dropdown to provide different filter capabilities is a pattern
we already use on the config-errors page.

This feature might be very useful for the "Watch my change as it
progresses through the system" use case.

If the admin actions are also visible, the search-plus icon will shift
to the left.

Change-Id: I3955b58440b81fab9a517ab8368a21901eb60783

SQUASH: FilterIcon

This feature is also very useful to fulfill the
"watch my change as it progresses through the system"
use case.

Change-Id: I28e9d7841fc05ce20702d7334cf8114ffdc20735