Add filter icon to QueueItem actions
On the new status page we don't have a link to the "Single Change Panel"
anymore as it doesn't provide more details compared to what's already
shown on the status page.

However, providing a filtered link that directly points to a single
QueueItem might be useful so that one can share a link to the change
with someone. To not confuse with any link or share functionality, we
just use a plain filter icon and add a tooltip.

If the admin actions are also visible, the filter icon will shift to the
left.

Change-Id: I3955b58440b81fab9a517ab8368a21901eb60783