Add share icon to QueueItem actions
On the new status page we don't have a link to the "Single Change Panel"
anymore as it doesn't provide more details compared to what's already
shown on the status page.

As one might still find it useful to be able to share the link to a
single change with someone, add a share icon to the QueueItem's actions.

If the admin actions are also visible, the share icon will shift to the
left.

Change-Id: Ic80954289822119388cca7e50b88651f95e9fc8e