Complete preliminary provider node lifecycle
This change implements the provider node lifecycle with the following
rules:

* in order to update a provider node a lock must be held
* as long as the nodeset request exists, the provider nodes will also
  be kept (though failed instance might already be cleaned up)
* if the nodeset request no longer exists the provider nodes must be
  locked (e.g. while in-use) to prevent cleanup
* an unlocked node without an associated request will be cleaned up

Eventually instead of deleting all unlocked nodes w/o a request, we
should check if the node could be re-used instead. However, the
implementation for that will be done at a later stage as this will have
some commonalities with the handling of 'min-ready' in the global
request processing phase.

Change-Id: Ia6524d608bacbb393262574857eead81a6467d25