Make ZK component more robust against ZK errors
For some Zookeeper related errors during update of a ZK component the
operation might still succeed despite an exception.

However, as we did not update the zstat in those cases, we could get
into a situation where locally tracked version diverged from the actual
znode version.

Since the component znode is emphemeral and only the owner should write
to that node, it should be safe to not supply the znode version during
update.

Components that might have the same name due to a misconfiguration
should already fail at the register stage when the emphemeral node is
created.

Change-Id: Ic06ca08f00a4660902873a4efddb220c5e8a0a14